+++ xc3_lib/src/msrd.rs
@@ -43,7 +43,7 @@ use crate::{
 };
 use bilge::prelude::*;
 use binrw::{args, binread, helpers::until_eof, BinRead, BinResult, BinWrite};
-use xc3_write::{write_full, Xc3Write, Xc3WriteOffsets};
+use xc3_write::{Xc3Write, Xc3WriteOffsets};
 
 /// Utilities for extracting and rebuilding streaming data.
 pub mod streaming;