+++ xc3_write/tests/count_offset.rs
@@ -1,7 +1,7 @@
 use std::io::Cursor;
 
 use hexlit::hex;
-use xc3_write::{assert_hex_eq, write_full, Endian, Xc3Write, Xc3WriteOffsets};
+use xc3_write::{assert_hex_eq, Endian, WriteFull, Xc3Write, Xc3WriteOffsets};
 
 #[test]
 fn write_count_offset() {
@@ -35,7 +35,9 @@ fn write_count_offset_full() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(04000000 08000000 01020304), writer.into_inner());
     assert_eq!(12, data_ptr);
@@ -55,7 +57,9 @@ fn write_count_offset_full_align_0x0() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(
         hex!(04000000 10000000 00000000 00000000 01020304),
@@ -78,7 +82,9 @@ fn write_count_offset_full_align_0xff() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(
         hex!(04000000 10000000 ffffffff ffffffff 01020304),