+++ xc3_write/tests/offset.rs
@@ -1,7 +1,7 @@
 use std::io::Cursor;
 
 use hexlit::hex;
-use xc3_write::{assert_hex_eq, write_full, Endian, Xc3Write, Xc3WriteOffsets};
+use xc3_write::{assert_hex_eq, Endian, WriteFull, Xc3Write, Xc3WriteOffsets};
 
 #[test]
 fn write_offset() {
@@ -39,7 +39,9 @@ fn write_offset_full() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(08000000 00000000 01000000), writer.into_inner());
     assert_eq!(12, data_ptr);
@@ -57,7 +59,9 @@ fn write_offset_full_align_0x0() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(08000000 00000000 01000000), writer.into_inner());
     assert_eq!(12, data_ptr);
@@ -75,7 +79,9 @@ fn write_offset_full_align_0xff() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(08000000 ffffffff 01000000), writer.into_inner());
     assert_eq!(12, data_ptr);
@@ -93,7 +99,9 @@ fn write_offset_full_optional_offset_some() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(04000000 01000000), writer.into_inner());
     assert_eq!(8, data_ptr);
@@ -112,7 +120,9 @@ fn write_offset_full_optional_offset_none() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(00000000), writer.into_inner());
     assert_eq!(4, data_ptr);