+++ xc3_write/tests/optional_field.rs
@@ -1,7 +1,7 @@
 use std::io::Cursor;
 
 use hexlit::hex;
-use xc3_write::{assert_hex_eq, write_full, Endian, Xc3Write, Xc3WriteOffsets};
+use xc3_write::{assert_hex_eq, Endian, WriteFull, Xc3Write, Xc3WriteOffsets};
 
 #[test]
 fn write_offset_full_some() {
@@ -16,7 +16,9 @@ fn write_offset_full_some() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(08000000 02000000 01000000), writer.into_inner());
     assert_eq!(12, data_ptr);
@@ -36,7 +38,9 @@ fn write_offset_full_none() {
 
     let mut writer = Cursor::new(Vec::new());
     let mut data_ptr = 0;
-    write_full(&value, &mut writer, 0, &mut data_ptr, Endian::Little, ()).unwrap();
+    value
+        .write_full(&mut writer, 0, &mut data_ptr, Endian::Little, ())
+        .unwrap();
 
     assert_hex_eq!(hex!(04000000 01000000), writer.into_inner());
     assert_eq!(8, data_ptr);